requirements.txt
psycopg2-binary
beautifulsoup4
requests
tqdm
mcp
python-dotenv